cleanup recording links
Some of the PBS-related recordings are very obsolete, and the link to them are removed (while still keeping the topic in the table for future reference). There are also other recordings which are still kept, and they are partly useful. E.g. storage, worker, modules, etc. We have to revise these, too, but for now they can still stay, because they can still convey the main message (when taken with a grain of salt).